% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_utils.R
\name{check_cols}
\alias{check_cols}
\title{Check for required columns}
\usage{
check_cols(user_df, col_df)
}
\arguments{
\item{user_df}{User-supplied dataframe.}

\item{col_df}{Dataframe of columns to check against.}
}
\value{
Error if required columns are missing, otherwise nothing.
}
\description{
Check for required columns
}
\keyword{internal}
