% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_FVS_files.R
\name{check_db}
\alias{check_db}
\title{Helper function to check database validity for running FVS}
\usage{
check_db(database)
}
\arguments{
\item{database}{Character string. File path to the SQLite input database.}
}
\value{
Nothing. Errors if FVS_StandInit and/or FVS_TreeInit not found in
the database.
}
\description{
Helper function to check database validity for running FVS
}
\keyword{internal}
