% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_FIA_tree_list.R
\name{clean_FIA_treeList}
\alias{clean_FIA_treeList}
\title{Clean an FIA tree list to prepare for FVS.}
\usage{
clean_FIA_treeList(treelist, standinfo)
}
\arguments{
\item{treelist}{Tree list, from FIA_TREEINIT_PLOT table (FVS_TreeInit from get_FIA())}

\item{standinfo}{Stand information associated with the tree list (FVS_StandInit from get_FIA())}
}
\value{
A dataframe containing all tree information necessary to be input into FVS.
}
\description{
Clean an FIA tree list to prepare for FVS.
}
