% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_trees}
\alias{ex_trees}
\title{Example Tree list}
\format{
\subsection{\code{ex_trees}}{

A data frame with 51 rows and 31 columns.
}
}
\usage{
ex_trees
}
\description{
A subset of FIA data for just one plot in western Montana.
These data come from the FVS_TREEINIT_PLOT table and have been processed using \code{\link[=get_FIA_state]{get_FIA_state()}}
with the option "add_identifiers" set to TRUE.
}
\keyword{datasets}
