% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_FVS_files.R
\name{fill_tree_list}
\alias{fill_tree_list}
\title{Fill missing values from tree list using stand information.}
\usage{
fill_tree_list(tree_list, stand_info)
}
\arguments{
\item{tree_list}{Tree list. Dataframe.}

\item{stand_info}{Stand information associated with the tree list. Dataframe.}
}
\value{
A dataframe containing all tree information necessary to be input
into FVS.
}
\description{
Fill missing values from tree list using stand information.
}
\keyword{internal}
