% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_habcode.R
\name{fmt_pv}
\alias{fmt_pv}
\title{Format PV codes to match lookup tables}
\usage{
fmt_pv(pv_code)
}
\arguments{
\item{pv_code}{}
}
\value{
Character string of 10 characters.
}
\description{
Format PV codes to match lookup tables
}
\keyword{internal}
