% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FIA.R
\name{get_FIA}
\alias{get_FIA}
\title{Get FVS-ready FIA data}
\usage{
get_FIA(db_loc, fia_cond_subset, verbose = FALSE)
}
\arguments{
\item{db_loc}{Location for the FIA database.}

\item{fia_cond_subset}{A subset of an FIA COND table.}

\item{verbose}{Boolean value. If TRUE, will print SQL queries to console.}
}
\value{
List of 2. FVS_StandInit is the stand table that can be used for
}
\description{
Get FVS-ready FIA data
}
