% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cond.R, R/get_FVS_ready.R
\name{fetch_cond}
\alias{fetch_cond}
\alias{get_FVS_ready}
\title{Fetch filtered FIA condition table for subsetting}
\usage{
fetch_cond(db_loc, filter_statements = "")

get_FVS_ready(
  database,
  fia_cond_subset,
  type = c("PLOT", "COND"),
  verbose = FALSE,
  add_identifiers = FALSE,
  to_db = TRUE,
  output_database = "fvs_ready.db"
)
}
\arguments{
\item{db_loc}{Character string. Location for the FIA database.}

\item{filter_statements}{Character string of dplyr-style filter statements.
If not provided, entire (large!) condition table will be returned. Column
names in the filter_statements arguments must match columns in the COND
table from the \href{https://research.fs.usda.gov/sites/default/files/2025-08/wo-v9-4_Aug2025_UG_FIADB_database_description_NFI.pdf}{NFI database}.
See examples.}

\item{database}{Character string. Location for the FIA database.}

\item{fia_cond_subset}{Dataframe. A subset of an FIA COND table; will be used
in an inner join to filter the FVS_StandInit and FVS_TreeInit tables.}

\item{type}{Character string, one of "PLOT" or "COND" (default "PLOT").
Specifies whether FVS_<Stand/Tree>Init_PLOT or FVS_<Stand/Tree>Init_COND
should be returned.}

\item{verbose}{Boolean value. If TRUE, will print SQL queries to console.}

\item{add_identifiers}{Boolean value. If TRUE, will add a PID (Plot
IDentifier) column to the stand table and a TUID (Tree Unique IDentifier)
column to the tree table. PID and TUID are unique, persistent identifiers.
They uniquely identify each FIA plot and each tree in each plot. Unlike
FIA-provided identifiers, these stay the same across all years.}

\item{to_db}{Boolean. If TRUE (default), will write the filtered dataframes
to an FVS-ready database that can be passed to \code{\link[=write_FVS_key]{write_FVS_key()}} or
\code{\link[=write_multistand_key]{write_multistand_key()}}.}

\item{output_database}{Character string. External location for filtered data
tables. Default is 'fvs_ready.db'}
}
\value{
\code{fetch_cond()}: COND data frame for use as a filter for
\code{get_FVS_ready()}.

\code{get_FVS_ready()}: List of 2. $FVS_StandInit is a dataframe of the
stand information. $FVS_TreeInit is a dataframe of all tree measurements. A
single stand from this list selected with STAND_CN and the associated tree
list (matching STAND_CN) can be passed to \code{run_FVS()}.
}
\description{
\code{fetch_cond()} is a convenience function for fetching subsets of COND tables
for use with \code{\link[=get_FVS_ready]{get_FVS_ready()}}.

\code{get_FVS_ready()} fetches FVS Stand and FVS Tree tables from a downloaded
state-level FIA database (from the \href{https://apps.fs.usda.gov/fia/datamart/datamart.html}{FIA datamart website}). Note that
these tables are only available at the state-level, so this function only
works with state-level databases.
}
\examples{

# get just the condition table:
database <- system.file('extdata', 'dummy_fia.db', package = 'rFVSIEtools')
fetch_cond(database, 'STATECD == 30, INVYR >= 2001')

# if filtering by a character column, use "" to surround filter statement and
#> '' to surround the string:
fetch_cond(database, "CN == '303115670489998'")

# fetch_cond() dataframe can go directly into get_FIA_state():

database <- system.file('extdata', 'dummy_fia.db', package = 'rFVSIEtools')
cond <- fetch_cond(database, 'STATECD == 30, INVYR >= 2001')
get_FVS_ready(database, cond, type = 'PLOT', to_db = FALSE)


}
