% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_habcode.R
\name{map_habcode}
\alias{map_habcode}
\title{Map PV/PV ref code pairs to associated FVS-IE habitat type}
\usage{
map_habcode(pv_code, pv_ref, verbose = FALSE)
}
\arguments{
\item{pv_code}{Potential vegetation code (scalar)}

\item{pv_ref}{Potential vegetation reference code (scalar)}

\item{verbose}{If TRUE, report when NAs are returned and why (no matches are
found or PV ref code is blank). Default FALSE.}
}
\value{
3-digit integer representing FVS-IE habitat type. See
\href{https://www.fs.usda.gov/sites/default/files/forest-management/fvs-ie-overview.pdf}{FVS-IE
Overview}
}
\description{
Note that this function is not (currently) vectorized, so
applying to a dataframe takes an extra step. See examples.
}
\details{
Note that this function relies on lookup tables created for internal
use. To see the source code for creating these tables, please see
\url{https://github.com/spracha/FVSIE-tools/data-raw/ref_tables.R}.
}
\examples{
# can be used for a single pair of values:
map_habcode('578', 110)
# <int> 570

# more realistically, you'll want to apply to a whole dataset:
habs <- data.frame(pv_code = c('578', 'CES313', '10'),
                   pv_ref = c(110, 626, 110))

# dplyr-style:
habs |> dplyr::rowwise() |> dplyr::mutate(hab_cd = map_habcode(pv_code, pv_ref))
# A tibble: 3x3
# pv_code    pv_ref     hab_cd
# <chr>      <dbl>      <int>
# 578        110        570
# CES313     626        720
# 10         110        130

# Base R:

habs$hab_cd <- unname(mapply(map_habcode, pv_code = habs$pv_code, pv_ref = habs$pv_ref))
habs
# a dataframe
# pv_code    pv_ref     hab_cd
# <chr>      <dbl>      <int>
# 578        110        570
# CES313     626        720
# 10         110        130
}
