% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_columns.R
\name{get_tree_columns}
\alias{get_tree_columns}
\alias{get_stand_columns}
\title{Get a table of needed tree and/or stand list columns}
\usage{
get_tree_columns()

get_stand_columns()
}
\value{
A dataframe (32 rows x 5 columns for tree column table, 58 rows x 5
columns for stand column table.).
}
\description{
\code{\link[=run_FVS]{run_FVS()}} and \code{\link[=run_FVS_parallel]{run_FVS_parallel()}} assumes that columns have
certain names. These assumed names come from FIA's FVS_TREE_INIT and
FVS_STAND_INIT tables, with the exception of \code{FVS_HAB} (the FVS habitat
type associated with the PV Code/PV Ref code combination, or obtained by
coercing PV code to an integer -- see \code{\link[=map_habcode]{map_habcode()}}). These functions
show you the default names for required columns (\code{col}), their default
value, if applicable (\code{default}; 9999 means so default), whether or not the
column has a default (\code{has_default}), acceptable alternative column names
(\code{alt_col}), and the required data type of the column (\code{dtype}).
}
\details{
From Essential FVS: A User's Guide to the Forest Vegetation
Simulator (Dixon, updated March 2026), p. 47
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
View(get_tree_columns())
\dontshow{\}) # examplesIf}
print(get_tree_columns())
print(get_stand_columns())

}
\seealso{
\code{\link[=set_tree_cols]{set_tree_cols()}}, \code{\link[=set_stand_cols]{set_stand_cols()}}
}
