% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_FVS_parallel.R
\name{run_FVS_parallel}
\alias{run_FVS_parallel}
\title{Run multiple stands through FVS, in parallel.}
\usage{
run_FVS_parallel(
  stand_info,
  tree_list,
  n_batches = 1,
  simple_output = TRUE,
  out_dir,
  fvs_bin,
  year_col = NULL,
  ...
)
}
\arguments{
\item{stand_info}{A dataframe where each row is a stand. Must contain
STAND_CN column.}

\item{tree_list}{A dataframe where each row is a tree record. Must contain
STAND_CN column. If the trees in this dataframe are not associated with a
stand in \code{stand_info}, or if no trees are alive, the stand will be skipped.}

\item{n_batches}{Optional. If a positive integer is supplied, the stands will
be grouped into \code{n_batches} groups. Each group will be run in parallel, but
groups will be run sequentially. This is faster than attempting to run all
stands in parallel when there are a large number of stands (>1000 or so).}

\item{simple_output}{Optional, default TRUE. See Values.}

\item{out_dir}{Directory to put .key, .tre, and .out files.}

\item{fvs_bin}{Location of FVS software.}

\item{year_col}{Optional character string. If additional years to report
varies by stand, this argument specifies the column in the stand
information dataframe that contains those additional reporting years.}

\item{...}{Additional parameters to pass on to control simulation. See
\code{\link[=write_FVS_files]{write_FVS_files()}} for details.}
}
\value{
If \code{simple_output} is \code{TRUE}, returns a list of two dataframes:
\code{treelist} and \code{summary}. These dataframes are the outputs from
\code{\link[=run_FVS]{run_FVS()}}, with all stands and trees bound together into a single
dataframe.

If \code{simple_output} is \code{FALSE}, returns a list of nested lists. Each nested
list contains both \code{treelist} and \code{summary} for a single stand from
\code{\link[=run_FVS]{run_FVS()}}.
}
\description{
This function uses the \code{future.apply} and \code{future} packages to support
parallel processing. Consult \code{future} package documentation for more details.
}
