% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_utils.R
\name{search_cols}
\alias{search_cols}
\title{Search a user-supplied dataframe for necessary columns}
\usage{
search_cols(user_df, col_df)
}
\arguments{
\item{user_df}{User-supplied dataframe}

\item{col_df}{Dataframe containing column information (column name,
alternative names, defaults)}
}
\value{
6-column dataframe:
\emph{required_col}: Character. Name of required column from col_df.
\emph{data_col}: Character. Name of the matching column found in user_df.
\emph{default}: Character. Default value for the column, 9999 if no default is available.
\emph{has_default}: Boolean. Does the column have a default value?
\emph{dtype}: Character. FVS required datatype for the default value.
}
\description{
Search a user-supplied dataframe for necessary columns
}
\keyword{internal}
