% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_columns.R
\name{set_default_values}
\alias{set_default_values}
\title{Using a dataframe of requirements and defaults, set default values where needed.}
\usage{
set_default_values(req_var, col_info, user_df, quiet)
}
\arguments{
\item{req_var}{Character. Required variable to set default for.}

\item{col_info}{Dataframe, from running search_cols().}

\item{user_df}{User-supplied data frame}

\item{quiet}{Boolean. Should defaults be set without messages (TRUE) or with
messages (FALSE)?}
}
\value{
Vector, length 1 if default filled in, length(nrow(data_col)) if no
default needed.
}
\description{
Using a dataframe of requirements and defaults, set default values where needed.
}
\keyword{internal}
