% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_FVS_files.R
\name{set_FVSie_defaults}
\alias{set_FVSie_defaults}
\title{Replace NA values in stand tables with FVS-IE defaults for writing a stand list.}
\usage{
set_FVSie_defaults(stand_info)
}
\arguments{
\item{stand_info}{Stand to set defaults for.}
}
\value{
A dataframe.
}
\description{
Replace NA values in stand tables with FVS-IE defaults for writing a stand list.
}
\keyword{internal}
