% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_FVS_parallel.R
\name{simplify_outputs}
\alias{simplify_outputs}
\title{Simplify nested lists of FVS outputs from running FVS in parallel.}
\usage{
simplify_outputs(output_lists)
}
\arguments{
\item{output_lists}{}
}
\value{
List of 2. \verb{$tree_list} is a list of all trees at the initial and
projected time(s), with one row for each tree. \verb{$summary} is the stand
summary; see \code{rFVS::fvsGetSummary()} for details.
}
\description{
Simplify nested lists of FVS outputs from running FVS in parallel.
}
\keyword{internal}
