% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_FVS_files.R
\name{write_FVS_files}
\alias{write_FVS_files}
\title{Write keyword files and treelist files.}
\usage{
write_FVS_files(
  stand_info,
  tree_list,
  out_dir,
  proj_len = 100,
  calibrate = TRUE,
  triple = FALSE,
  add_regen = FALSE,
  custom_SDI_max = NULL,
  random_seed = NULL,
  STDIDENT = "FVSProjection",
  file_prefix = NULL,
  ...,
  additionals = NULL,
  estab = NULL
)
}
\arguments{
\item{stand_info}{Stand data to be run through FVS, e.g., FVS_StandInit
output from get_FIA()}

\item{tree_list}{Tree list to be run through FVS, e.g., FVS_TreeInit output
from get_FIA()}

\item{out_dir}{Directory to write files to.}

\item{proj_len}{How many years ahead should we project? Default 100.}

\item{calibrate}{Logical. Should self-calibration be used? Default TRUE.}

\item{triple}{Logical. Should tripling be turned on? Default FALSE.}

\item{add_regen}{Logical. Should regeneration be modeled? Default FALSE.}

\item{custom_SDI_max}{Optional. Maximum stand density index.}

\item{random_seed}{For replicability. Default 2025. To turn off, set to NULL.}

\item{STDIDENT}{Optional. Stand Identity keyword to pass to FVS.}

\item{file_prefix}{Optional. If provided, KEY and TRE files with temp names
will have this prefix.}

\item{...}{Additional optional arguments passed to \code{\link[=write_FVS_KEY]{write_FVS_KEY()}}. See
details.}

\item{additionals}{Additional lines (keywords + fields) to print to keyword
file. These must be formatted according to FVS keyword documentation. See
examples.}

\item{estab}{Additional lines (keyword + fields) to print to keyword file
following the ESTAB keyword. These must be formatted according to FVS
keyword documentation.}
}
\value{
The filename created for .key and .tre files, invisibly.
}
\description{
This function, wrapped in \code{\link[=run_FVS]{run_FVS()}}, writes and saves .key and .tre files
that FVS-IE uses to initialize and carry out a simulation. Use outside of
\code{run_FVS()} when diagnostics on these files are needed.
}
\details{
Optional arguments:
\itemize{
\item \code{custom_SDI_max}: Dataframe. \code{custom_SDI_max$SP} is the species, \verb{ }Custom_SDI_max$MaxSDI` is that species' SDI.
\item \code{TIMEINT}: Named list. \code{TIMEINT$CYCLE_NUM} is the cycle whose length is to
be changed. 0 changes the length for all cycles. \code{TIMEINT$CYCLE_LEN} is the
cycle length to change to.
\item \code{READCORD}, \code{READCORR}: Vectors of numeric values. Length must be 23 to
match the number of species in the variant.
\item \code{CYCLEAT}: Additional reporting year.
}
}
\examples{
clean_stand <- set_stand_cols(ex_stand)
clean_trees <- set_tree_cols(ex_trees, stand_info = clean_stand)
out_dir <- tempdir()

# A run with all defaults: 100 years, self-calibration, no tripling, no autoestablishment
write_FVS_files(clean_stand, clean_trees, out_dir)

# A run with carbon reporting requested using the keyword sequence:
# FMIN
# CARBREPT
# END
# Carbon results will also be automatically saved to the output database.

carb_key <- paste0('\nFMIN\n','CARBREPT','\n', 'END\n')
write_FVS_files(clean_stand, clean_trees, out_dir, additionals = carb_key)

# clean-up temp files
file.remove(list.files(out_dir))
unlink(out_dir)
}
