% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_FVS_files.R
\name{write_FVS_KEY}
\alias{write_FVS_KEY}
\title{Write FVS Keyword File}
\usage{
write_FVS_KEY(
  stand,
  proj_len,
  calibrate,
  triple,
  add_regen,
  custom_SDI_max,
  random_seed,
  STDIDENT,
  ...,
  additionals = NULL,
  keyfile_name
)
}
\arguments{
\item{stand}{Dataframe. Stand data (for a single stand) for running FVS.}

\item{proj_len}{How long, in years, should the projection be? Default 100.}

\item{calibrate}{Logical. Should self-calibration be used? Default TRUE.}

\item{triple}{Logical. Should tripling be turned on? Default FALSE.}

\item{add_regen}{Logical. Should regeneration be modeled? Default FALSE.}

\item{custom_SDI_max}{Optional. Maximum stand density index. See details.}

\item{random_seed}{For replicability. Default 2025. To turn off, set to NULL.}

\item{STDIDENT}{Optional. Stand Identity keyword to pass to FVS.}

\item{...}{Optional arguments. See details.}

\item{additionals}{Additional lines to print to keyword file. These must be
formatted according to FVS keyword documentation.}

\item{keyfile_name}{.key file to write to.}
}
\value{
Keyword filename, invisibly.
}
\description{
Given a stand list and FVS projection parameters, writes a keyword file. This
can then be inspected and passed to run_FVS().
}
\details{
Optional arguments:
\itemize{
\item \code{custom_SDI_max}: Dataframe. custom_SDI_max$SP is the species,
\code{custom_SDI_max$MaxSDI} is that species' SDI.
\item \code{TIMEINT}: Named list. TIMEINT$CYCLE_NUM is the cycle whose length is to
be changed. 0 changes the length for all cycles. TIMEINT$CYCLE_LEN is the
cycle length to change to.
\item \code{READCORD}, \code{READCORR}: Vectors of numeric values. Length must be 23 to
match the number of species in the variant, and the order of species must
match FVS-IE's species sequence order.
\item \code{CYCLEAT}: Additional reporting year.
}
}
\keyword{internal}
