% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_utils.R
\name{check_name}
\alias{check_name}
\title{A convenience function for column name-checking.}
\usage{
check_name(name, check_against)
}
\arguments{
\item{name}{String to check}

\item{check_against}{Pattern to check against.}
}
\value{
A single Boolean value.
}
\description{
A convenience function for column name-checking.
}
\keyword{internal}
