% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_stand}
\alias{ex_stand}
\title{Example Stand Information Table}
\format{
\subsection{\code{ex_stand}}{

A data frame with 1 rows and 34 columns.
}
}
\usage{
ex_stand
}
\description{
A subset of FIA data for just one plot in western Montana.
These data come from the FVS_STANDINIT_PLOT table and have been processed using \code{\link[=get_FIA_state]{get_FIA_state()}}
with the option "add_identifiers" set to TRUE.
}
\keyword{datasets}
