% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_keyword.R
\name{format_keyword}
\alias{format_keyword}
\title{Format an FVS keyword}
\usage{
format_keyword(keyword, ..., dtypes = NULL, round_to = 2)
}
\arguments{
\item{keyword}{A character string. The FVS keyword name.}

\item{...}{Optional values for fields associated with the keyword. Will be
coerced to the type specified by \code{dtypes} if specified or character if
\code{dtypes} is not specified.}

\item{dtypes}{Optional character vector specifying the datatype of each
field. The length needs to match the number of fields. 's' for string, 'i'
for integer, and 'f' for float (decimals)}

\item{round_to}{An optional integer specifying how many places floats should
be rounded to (default 2).}
}
\value{
A character string.
}
\description{
Add whitespace and use optional datatypes to format an FVS keyword to match
requirements. This function does not format READCOR* values, which must be
passed to \code{\link[=write_FVS_key]{write_FVS_key()}} as vectors in order of the species in the
variant.
}
\examples{
format_keyword('NOINGROW')

format_keyword('SDICALC', 0, 1, 1)

# Lower case will be converted to upper case,
#  and any spaces will be removed in the keyword:
format_keyword('sdi calc', 0, 1, 1)

# not recommended to use for floats without specifying datatype:
# BAD
try(format_keyword('SDICALC', 1/3, 1, 1))

# OK
format_keyword('SDICALC', dtypes = c('f', 'i', 'i'), 1/3, 1, 1)

# For multiple keywords to go one after another, use paste() with sep = '\n':
additionals <- paste(format_keyword('SDICALC', dtypes = c('f', 'i', 'i'), 1/3, 1, 1),
                     format_keyword('SDIMAX',
                                    dtypes = c('i', 'f', 's', 's', 'f', 'f', 'i'),
                                    1, 300, '', '', 65, 90, 0),
                     sep = "\n")
cat(additionals)

}
