% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FIA_state.R
\name{get_FIA_state}
\alias{get_FIA_state}
\title{Get FVS-ready FIA data from a state-level database}
\usage{
get_FIA_state(db_loc, fia_cond_subset, verbose = FALSE)
}
\arguments{
\item{db_loc}{Location for the FIA database.}

\item{fia_cond_subset}{Dataframe. A subset of an FIA COND table.}

\item{verbose}{Boolean value. If TRUE, will print SQL queries to console.}
}
\value{
List of 2. FVS_StandInit is the stand table that can be used for
}
\description{
\code{get_FIA_state()} fetches FVS Stand and FVS Tree tables from a downloaded FIA
database. Note that these tables are only available at the state-level, so
this function only works with state-level databases.
}
\examples{

db_loc <- system.file('extdata', 'dummy_fia.db', package = 'rFVSIEtools')
con <- DBI::dbConnect(RSQLite::SQLite(), db_loc)
cond <- dplyr::tbl(con, 'COND') |>
        dplyr::filter(STATECD == 30) |>
        dplyr::collect()
DBI::dbDisconnect(con)
get_FIA_state(db_loc, cond)

}
