% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_columns.R
\name{get_tree_columns}
\alias{get_tree_columns}
\alias{get_stand_columns}
\title{Get a table of needed tree and/or stand list columns}
\usage{
get_tree_columns()

get_stand_columns()
}
\value{
A dataframe (32 rows x 5 columns for tree column table, 58 rows x 5
columns for stand column table.).
}
\description{
These functions show you the default names for required columns
(\code{col}), their default value, if applicable (\code{default}; 9999 means so
default), whether or not the column has a default (\code{has_default}),
acceptable alternative column names (\code{alt_col}), and the required data type
of the column (\code{dtype}).
}
\details{
From Essential FVS: A User's Guide to the Forest Vegetation
Simulator (Dixon, updated March 2026), p. 47
}
\examples{
print(get_tree_columns())
print(get_stand_columns())


}
