% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_FVS.R
\name{run_FVS}
\alias{run_FVS}
\title{Run FVS from a database with stand and tree information.}
\usage{
run_FVS(
  fvs_bin,
  variant = c("ie", "ak", "bm", "ca", "ci", "cr", "cs", "ec", "em", "nc", "kt", "ls",
    "ne", "pn", "sn", "so", "tt", "ut", "wc", "ws"),
  keyword_file,
  stop_year = NULL,
  stop_code = NULL,
  stop_file = NULL,
  verbose = FALSE
)
}
\arguments{
\item{fvs_bin}{Character string. The file path where FVS DLLs are located.}

\item{variant}{Character string. The two-letter code corresponding to an FVS
variant. Default is Inland Empire ('ie').}

\item{keyword_file}{Character string. Path to the keyword file controlling
the simulation; include the .key extension. See \code{\link[=write_FVS_key]{write_FVS_key()}} for
keyword file generation.}

\item{stop_year}{Optional integer. Year to stop simulation and store
simulation state.}

\item{stop_code}{Optional integer. State of the simulation to store.}

\item{stop_file}{Optional character string. Where to store the stand
variables at the supplied stop point.}

\item{verbose}{Logical. Should FVS library loading/unloading be accompanied
by a message? Default \code{FALSE}.}
}
\value{
The FVS program name (e.g., 'FVSie'), invisibly. Simulation outputs
are sent to the file specified in the keyword file and to a '.out' file
(location and prefix specified in keyword file).
}
\description{
Run FVS from a database with stand and tree information.
}
\examples{
\dontrun{
out_dir <- tempdir()
database <- system.file(file.path('extdata', 'ex_data.db'), package = 'rFVSIEtools')
kwd <- write_FVS_key(out_dir = out_dir, file_prefix = 'example_kwd', database = database)

run_FVS(fvs_bin = 'C:/FVS/FVSSoftware/FVSbin', variant = 'ie', keyword_file = kwd)
}

}
