% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_FVS.R
\name{run_FVS}
\alias{run_FVS}
\title{Run FVS-IE.}
\usage{
run_FVS(trees, standinfo, outdir, fvs_bin, ..., verbose = FALSE)
}
\arguments{
\item{trees}{Tree list, e.g., from get_FIA_state().}

\item{standinfo}{Stand information for the single stand corresponding to \code{treelist}.}

\item{outdir}{Directory to write keyword, tree, and .out files to.}

\item{fvs_bin}{FVS software location.}

\item{...}{Additional arguments passed to write_FVS_files() to control simulation.}

\item{verbose}{If TRUE, report the names of the summary table and the number of rows in each year.}
}
\value{
A list of two. $treelist is the combined tree list from year 0 and the final simulation year; summary is the FVS summary table.
}
\description{
Run FVS-IE.
}
