% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_checking.R
\name{search_cols}
\alias{search_cols}
\title{Search a user-supplied dataframe for necessary columns}
\usage{
search_cols(user_df, col_df)
}
\arguments{
\item{user_df}{User-supplied dataframe}

\item{col_df}{Dataframe containing column information (column name,
alternative names, defaults)}
}
\value{
2-column dataframe. First column is the required column, second
column is the matching column found in the data.
}
\description{
TODO: refactor to use matrix multiplication for speed improvements.
}
\keyword{internal}
