% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_checking.R
\name{set_stand_cols}
\alias{set_stand_cols}
\title{Set stand columns to use default names and values.}
\usage{
set_stand_cols(stand_info, quiet = TRUE)
}
\arguments{
\item{stand_info}{Stand information dataframe to clean.}

\item{quiet}{Boolean. Should defaults be set without messages (TRUE) or with
messages (FALSE)?}
}
\value{
Stand information data frame with all columns that FVS needs to run.
}
\description{
Set stand columns to use default names and values.
}
