% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_columns.R
\name{set_stand_cols}
\alias{set_stand_cols}
\title{Set stand columns to use default names and values.}
\usage{
set_stand_cols(stand_info, map_habcode = TRUE, map_verbose = FALSE)
}
\arguments{
\item{stand_info}{Stand information dataframe to clean.}

\item{map_habcode}{Boolean. Should PV Code/PV reference code pairs be mapped
to FVS habitat type codes? Default TRUE. If this is turned off, PV codes
will be coerced to integers, so plant association codes ('CXXXX') will be
filled with NA and treated as default habitat type (260).}

\item{map_verbose}{Boolean. Should habitat type mapping be done verbosely?
Default FALSE.}
}
\value{
Stand information data frame with all columns that FVS needs to
run.This dataframe has at least 58 columns (# of required stand columns) and
the same number of rows as the input.
}
\description{
Set stand columns to use default names and values.
}
