% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_checking.R
\name{set_tree_cols}
\alias{set_tree_cols}
\title{Set tree columns to use default names and values.}
\usage{
set_tree_cols(tree_list, quiet = TRUE)
}
\arguments{
\item{tree_list}{Tree list to clean.}

\item{quiet}{Boolean. Should defaults be set without messages (TRUE) or with
messages (FALSE)?}
}
\value{
Tree list data frame containing all required columns for FVS to run.
}
\description{
Set tree columns to use default names and values.
}
