% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_columns.R
\name{set_tree_cols}
\alias{set_tree_cols}
\title{Set tree columns to use default names and fill in default values}
\usage{
set_tree_cols(
  tree_list,
  stand_info = NULL,
  map_habcode = TRUE,
  map_verbose = FALSE
)
}
\arguments{
\item{tree_list}{Dataframe. Tree list to set names/values for.}

\item{stand_info}{(Optional) Dataframe. Stand information for the tree list.
If supplied (recommended), used to fill missing values for SLOPE, ASPECT,
PV_CODE, and PV_REF_CODE.}

\item{map_habcode}{Boolean. Should PV Code/PV reference code pairs be mapped
to FVS habitat type codes? Default TRUE. If this is turned off, PV codes
will be coerced to integers, so plant association codes ('CXXXX') will be
filled with NA and treated as default habitat type (260).}

\item{map_verbose}{Boolean. Should habitat type mapping be done verbosely?
Default FALSE.}
}
\value{
Tree list dataframe containing all required columns for FVS to run.
Returned dataframe has at least 32 columns (the number of required columns)
and the same number of rows as input.
}
\description{
Small differences in column naming conventions can cause FVS to
fail due to "column not found" errors. This function searches input tree
list for the necessary columns, with some alternative names possible. If
necessary columns aren't found, it attempts to fill in default values. If
the needed column is missing and there is no default value, this function
errors. See \code{\link[=get_tree_columns]{get_tree_columns()}} for a list of all required FVS columns.
}
