% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_FVS_files.R
\name{write_FVS_files}
\alias{write_FVS_files}
\title{Write keyword files and treelist files.}
\usage{
write_FVS_files(
  tree_list,
  stand_info,
  out_dir,
  proj_len = 100,
  calibrate = TRUE,
  triple = FALSE,
  add_regen = FALSE,
  custom_SDI_max = NULL,
  random_seed = NULL,
  STDIDENT = "FVSProjection",
  file_prefix = NULL,
  ...
)
}
\arguments{
\item{tree_list}{Tree list to be run through FVS, e.g., FVS_TreeInit output
from get_FIA()}

\item{stand_info}{Stand data to be run through FVS, e.g., FVS_StandInit
output from get_FIA()}

\item{out_dir}{Directory to write files to.}

\item{proj_len}{How many years ahead should we project? Default 100.}

\item{calibrate}{Logical. Should self-calibration be used? Default TRUE.}

\item{triple}{Logical. Should tripling be turned on? Default FALSE.}

\item{add_regen}{Logical. Should regeneration be modeled? Default FALSE.}

\item{custom_SDI_max}{Optional. Maximum stand density index.}

\item{random_seed}{For replicability. Default 2025. To turn off, set to NULL.}

\item{STDIDENT}{Optional. Stand Identity keyword to pass to FVS.}

\item{file_prefix}{Optional. If provided, KEY and TRE files with temp names
will have this prefix.}

\item{...}{Additional optional arguments passed to \code{\link[=write_FVS_KEY]{write_FVS_KEY()}}. See
details.}
}
\value{
The filename created for .key and .tre files, invisibly.
}
\description{
This function, wrapped in \code{\link[=run_FVS]{run_FVS()}}, writes and saves .key and .tre files
that FVS-IE uses to initialize and carry out a simulation. Use outside of
\code{run_FVS()} when diagnostics on these files are needed. If necessary columns
aren't found in the data, defaults will be filled verbosely; wrap the call in
\code{suppressMessages()} to ignore messages.
}
\details{
Optional arguments:
\itemize{
\item \code{custom_SDI_max}: Dataframe. \code{custom_SDI_max$SP} is the species, \verb{ }Custom_SDI_max$MaxSDI` is that species' SDI.
\item \code{TIMEINT}: Named list. \code{TIMEINT$CYCLE_NUM} is the cycle whose length is to
be changed. 0 changes the length for all cycles. \code{TIMEINT$CYCLE_LEN} is the
cycle length to change to.
\item \code{READCORD}, \code{READCORR}: Vectors of numeric values. Length must be 23 to
match the number of species in the variant.
\item \code{CYCLEAT}: Additional reporting year.
}
}
