% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_FVS_files.R
\name{write_FVS_files}
\alias{write_FVS_files}
\title{Writes keyword and treelist functions to files for running FVS.}
\usage{
write_FVS_files(
  treelist,
  standinfo,
  years_out = 100,
  calibrate = TRUE,
  triple = FALSE,
  add_regen = FALSE,
  customSDImax = NULL,
  randomseed = 2025,
  STDIDENT = "FVSProjection",
  outdir,
  file_prefix = NULL,
  ...
)
}
\arguments{
\item{treelist}{Tree list to be run through FVS, e.g., FVS_TreeInit output from get_FIA()}

\item{standinfo}{Stand data to be run through FVS, e.g., FVS_StandInit output from get_FIA()}

\item{years_out}{How many years ahead should we project?}

\item{calibrate}{Logical. Should self-calibration be used? Default TRUE.}

\item{triple}{Logical. Should tripling be turned on? Default FALSE.}

\item{add_regen}{Logical. Should regeneration be modeled? Default FALSE.}

\item{customSDImax}{Optional. Maximum stand density index.}

\item{randomseed}{For replicability. Default 2025. To turn off, set to NULL.}

\item{STDIDENT}{Optional. Stand Identity keyword to pass to FVS.}

\item{outdir}{Directory to write files to.}

\item{file_prefix}{Optional. If provided, KEY and TRE files with temp names will have this prefix.}

\item{...}{Additional optional arguments passed to write_FVS_KEY(). See details.}
}
\value{
The filename created for .key and .TRE files, invisibly.
}
\description{
Writes keyword and treelist functions to files for running FVS.
}
\details{
Optional arguments:
\itemize{
\item \code{customSDImax}: Dataframe. customSDImax$SP is the species, customSDImax$MaxSDI is that species' SDI.
\item \code{TIMEINT}: Named list. TIMEINT$CYCLE_NUM is the cycle whose length is to be changed. 0 changes the length for all cycles. TIMEINT$CYCLE_LEN is the
cycle length to change to.
\item \code{READCORD}, \code{READCORR}: Vectors of numeric values. Length must be 23 to match the number of species in the variant.
}
}
