% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_FVS_files.R
\name{write_FVS_key}
\alias{write_FVS_key}
\title{Write an FVS Keyword file.}
\usage{
write_FVS_key(
  out_dir = getwd(),
  file_prefix,
  database,
  STDIDENT = NULL,
  n_years = 100,
  CYCLEAT = NULL,
  calibrate = TRUE,
  htd_reg = TRUE,
  triple = TRUE,
  add_regen = TRUE,
  carbon_report = FALSE,
  estab_keywords = NULL,
  ffe_keywords = NULL,
  random_seed = NULL,
  db_tables = NULL,
  READCORD = NULL,
  READCORH = NULL,
  READCORR = NULL,
  additionals = NULL
)
}
\arguments{
\item{out_dir}{A character string specifying the directory for the keyword
file and FVS output. Defaults to working directory.}

\item{file_prefix}{A character string giving the keyword file name to go
before the '.key' extension.}

\item{database}{A character string specifying the full file path of the input
SQLite database.}

\item{STDIDENT}{Optional character string specifying the stand identity.
Should be NULL (default) for multistand projections; see \emph{THIS WOULD BE A
GOOD VIGNETTE}}

\item{n_years}{Integer specifying the number of years into the future to
project. Default 100.}

\item{CYCLEAT}{Optional integer specifying additional years to add cycles.
Useful for getting stand/tree information in specific years.}

\item{calibrate}{Logical. Should self-calibration be on (\code{TRUE}, default) or
turned off for the simulation (\code{FALSE})?}

\item{htd_reg}{Logical. Should missing heights be dubbed using heights in the
data (\code{TRUE}, default) or using regional defaults (\code{FALSE})?}

\item{triple}{Logical. Should tree records be split into three identical
records whose TPA sums to the original tree TPA in order to stabilize
estimates in small stands (\code{TRUE}, default) or not (\code{FALSE})?}

\item{add_regen}{Logical. Should the regeneration/establishment model
associated with the variant be used (\code{TRUE}, default) or turned off
(\code{FALSE})?}

\item{carbon_report}{Logical. Should the Fire and Fuels extension carbon
report be generated (\code{TRUE}) or not (\code{FALSE}, default)?}

\item{estab_keywords}{Optional character vector of keywords to control the
regeneration/establishment model (see Essential FVS 5.4). Each element in
the vector (or string if only one keyword is passed) must be formatted
according to the FVS keyword guide.}

\item{ffe_keywords}{Optional character vector of keywords to pass to the Fire
and Fuels Extension. Each element in the vector (or string if only one
keyword is passed) must be formatted according to the FVS keyword guide.}

\item{random_seed}{Optional odd integer to reseed the random number generator
for the base FVS model. Useful for reproducible simulation results.}

\item{db_tables}{Optional character vector of keywords to request additional
output tables.}

\item{READCORD}{Optional numeric vector of multipliers for the large tree
diameter growth model. The length and order of the vector must correspond
to the species in the FVS variant in order for multipliers to apply
correctly. See Essential FVS 6.5.2.2.}

\item{READCORH}{Optional numeric vector of multipliers for the large tree
height growth model. The length and order of the vector must correspond to
the species in the FVS variant in order for multipliers to apply correctly.
See Essential FVS 6.5.2.2.}

\item{READCORR}{Optional numeric vector of multipliers for the small tree
height growth model. The length and order of the vector must correspond to
the species in the FVS variant in order for multipliers to apply correctly.
See Essential FVS 6.5.2.2.}

\item{additionals}{Optional character vector of additional keywords. Each
element must be formatted according to the FVS keyword guide.}
}
\value{
The keyword file name, invisibly.
}
\description{
Given simulation parameters and file paths, this function writes
a .key file that can be used to control an FVS simulation. Automatic
formatting is supported for a variety of keywords; additional keywords may
be passed but must be formatted according to the \href{https://www.fs.usda.gov/sites/default/files/fvs-keyword.pdf}{FVS Keyword guide}.
Default simulation parameters (cycle length, number of years, calibration,
etc) are designed to match the FVS GUI defaults specified in Essential FVS,
except outputs are automatically written to a SQLite database.

By default, the keyword file specifies a simulation where:
\itemize{
\item Simulation results are output to a SQLite database called FVSOut.db in the working directory.
\item Cycles are 10 years in length.
\item The simulation is run for 100 years total.
\item Self-calibration is turned ON (see \href{https://www.fs.usda.gov/sites/default/files/forest-management/essential-fvs.pdf}{Essential FVS} section 6.4 for details).
\item Missing heights are dubbed using the height-diameter relationships from the input data.
\item Tripling is turned ON (see \href{https://www.fs.usda.gov/sites/default/files/forest-management/essential-fvs.pdf}{Essential FVS} p. 156 for details).
\item The regeneration establishment model is turned ON (see \href{https://www.fs.usda.gov/sites/default/files/forest-management/essential-fvs.pdf}{Essential FVS} section 5.4 for details).
}
}
\examples{
\dontrun{
  out_dir <- tempdir()
  database <- system.file(file.path('extdata', 'ex_data.db'), package = 'rFVSIEtools')
  write_FVS_key(out_dir = out_dir, file_prefix = 'example_kwd', database = database)

}
}
\seealso{
\code{\link[=format_keyword]{format_keyword()}} for formatting keywords; \code{\link[=write_multistand_key]{write_multistand_key()}}
for writing a keyword file to run multiple stands with the same simulation
parameters.
}
