% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_FVS_files.R
\name{write_FVS_KEY}
\alias{write_FVS_KEY}
\title{Write FVS Keyword File}
\usage{
write_FVS_KEY(
  stand,
  years_out = 100,
  calibrate = TRUE,
  triple = FALSE,
  add_regen = FALSE,
  customSDImax = NULL,
  randomseed = 2025,
  STDIDENT = NULL,
  ...,
  keyfilename
)
}
\arguments{
\item{stand}{Dataframe. Stand data (for a single stand) for running FVS.}

\item{years_out}{How many years out should we project? Default 100.}

\item{calibrate}{Logical. Should self-calibration be used? Default TRUE.}

\item{triple}{Logical. Should tripling be turned on? Default FALSE.}

\item{add_regen}{Logical. Should regeneration be modeled? Default FALSE.}

\item{customSDImax}{Optional. Maximum stand density index. See details.}

\item{randomseed}{For replicability. Default 2025. To turn off, set to NULL.}

\item{STDIDENT}{Optional. Stand Identity keyword to pass to FVS.}

\item{...}{Optional arguments. See details.}

\item{keyfilename}{.key file to write to.}
}
\value{
Keyword filename, invisibly.
}
\description{
Given a stand list and FVS projection parameters, writes a keyword file.
This can then be inspected and passed to run_FVS().
}
\details{
Optional arguments:
\itemize{
\item \code{customSDImax}: Dataframe. customSDImax$SP is the species, customSDImax$MaxSDI is that species' SDI.
\item \code{TIMEINT}: Named list. TIMEINT$CYCLE_NUM is the cycle whose length is to be changed. 0 changes the length for all cycles. TIMEINT$CYCLE_LEN is the
cycle length to change to.
\item \code{READCORD}, \code{READCORR}: Vectors of numeric values. Length must be 23 to match the number of species in the variant.
}
}
\keyword{internal}
