% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_FVS_files.R
\name{write_FVS_TRE}
\alias{write_FVS_TRE}
\title{Write a tree list dataframe to a .TRE file.}
\usage{
write_FVS_TRE(tree_list, treefile_name)
}
\arguments{
\item{tree_list}{Dataframe; tree list to write.}

\item{treefile_name}{Character. Destination for .TRE file.}
}
\value{
The tree file name, invisibly.
}
\description{
Given a tree list, write information to a properly formatted
.TRE file for FVS-IE to use.
}
\keyword{internal}
