% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_FVS_files.R
\name{write_FVS_TRE}
\alias{write_FVS_TRE}
\title{Writes tree list to .TRE file.}
\usage{
write_FVS_TRE(treelist, standinfo, treefilename)
}
\arguments{
\item{treelist}{Tree list to write.}

\item{standinfo}{Stand table associated with tree list.}

\item{treefilename}{Destination for .TRE file.}
}
\value{
The tree file name, invisibly.
}
\description{
Writes tree list to .TRE file.
}
\keyword{internal}
