% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_FVS_files.R
\name{write_FVS_TRE}
\alias{write_FVS_TRE}
\title{Writes tree list to .TRE file.}
\usage{
write_FVS_TRE(tree_list, stand_info, treefile_name)
}
\arguments{
\item{tree_list}{Tree list to write.}

\item{stand_info}{Stand table associated with tree list.}

\item{treefile_name}{Destination for .TRE file.}
}
\value{
The tree file name, invisibly.
}
\description{
Writes tree list to .TRE file.
}
\keyword{internal}
