% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_FVS_files.R
\name{write_multistand_key}
\alias{write_multistand_key}
\title{Write a keyword file to run multiple stands with the same simulation
parameters}
\usage{
write_multistand_key(
  STDIDENTs = NULL,
  out_dir = getwd(),
  database,
  file_prefix = "all_stands",
  addfile_ref = 40,
  ...
)
}
\arguments{
\item{STDIDENTs}{Optional character vector of stand identifiers specifying
the stands to be run in the simulation. If not provided, all STAND_ID
values in the stand table of the provided database will be used.}

\item{out_dir}{A character string specifying the directory for the keyword
file and FVS output. Defaults to working directory.}

\item{database}{A character string specifying the full file path of the input
SQLite database.}

\item{file_prefix}{An optional character string giving the keyword file name
to go before the '.key' extension. Default for multistand runs is
'all_stands'.}

\item{addfile_ref}{Optional integer file reference number, for FVS use. See
FVS keyword documentation for the ADDFILE keyword. Must be at least 31;
default is 40.}

\item{...}{Parameters passed to \code{\link[=write_FVS_key]{write_FVS_key()}}. See details.}
}
\value{
The name of the keyword file that runs all requested stands,
invisibly.
}
\description{
This function writes two .key files: one that defines the
keywords shared by all simulations ('shared_kwds.key') and one that tells
FVS to use these keywords to run the stands specified in \code{STDIDENTs} (or
all stands in the input database; '<file_prefix>.key'). Default shared
simulation parameters match those in the FVS GUI except outputs are
automatically written to an external file.

By default, the shared keyword file specifies a simulation where:
\itemize{
\item Simulation results are output to a SQLite database called FVSOut.db in the working directory.
\item Cycles are 10 years in length.
\item The simulation is run for 100 years total.
\item Self-calibration is turned ON (see \href{https://www.fs.usda.gov/sites/default/files/forest-management/essential-fvs.pdf}{Essential FVS} section 6.4 for details).
\item Missing heights are dubbed using the height-diameter relationships from the input data.
\item Tripling is turned ON (see \href{https://www.fs.usda.gov/sites/default/files/forest-management/essential-fvs.pdf}{Essential FVS} p. 156 for details).
\item The regeneration establishment model is turned ON (see \href{https://www.fs.usda.gov/sites/default/files/forest-management/essential-fvs.pdf}{Essential FVS} section 5.4 for details).
}

See details for simulation options.
}
\details{
The following named arguments can be passed to the \code{...} argument (see
\code{\link[=write_FVS_key]{write_FVS_key()}} for details):
\itemize{
\item \code{output_fmt}: A character string. Default \code{'SQLite'}..
\item \code{n_years}: Integer, default 100.
\item \code{CYCLEAT}: Optional integer.
\item \code{calibrate}: Logical, default \code{TRUE}.
\item \code{htd_reg}: Logical, default \code{TRUE}.
\item \code{triple}: Logical, default \code{TRUE}.
\item \code{add_regen}:  Logical, default \code{TRUE}.
\item \code{carbon_report}: Logical, default \code{FALSE}.
\item \code{estab_keywords}: Optional character vector of keywords.
\item \code{ffe_keywords}: Optional character vector of keywords.
\item \code{random_seed}: Optional odd integer.
\item \code{db_tables}: Optional character vector.
\item \code{READCORD}, \code{READCORH}, \code{READCORR}: Optional numeric vectors.
\item \code{READCORH}: Optional numeric vector.
\item \code{additionals}: Optional character vector.
}
}
\seealso{
\code{\link[=write_FVS_key]{write_FVS_key()}}
}
